Initial implementation of a 'generate' face for Puppet.
This commit represents an initial implementation for a 'generate' Puppet face.
The 'generate' face is responsible for generating Puppet source code from Ruby
implementations.

Initially this face only supports the 'types' action to generate resource types
in Puppet language using the 'pcore' format.

The 'pcore' format is under development and subject to change.